Create Sonar Predictive System
This project implements a Sonar-based predictive system using Logistic Regression to classify objects as either rock or mine. The dataset used contains sonar readings that can help distinguish between the two classes. This system was developed in Google Colab and is designed to help identify the type of object based on sonar data.